fix: :bug: update runtime config to new structure
